feat(Pagination): decouple the Pagination Component to separate package - in order to use this Pagination Component in other framework (like Aurelia-Slickgrid) without rewriting the same code, we can simply extract this into a brand new Pagination Component
