+++ packages/pagination-component/src/__tests__/slick-pagination.spec.ts
@@ -1,8 +1,8 @@
 
 import { PaginationService, SharedService, SlickGrid, GridOption } from '@slickgrid-universal/common';
+import { TranslateServiceStub } from '../../../../test/translateServiceStub';
+import { EventPubSubService } from '../../../vanilla-bundle/src/services/eventPubSub.service';
 import { SlickPaginationComponent } from '../slick-pagination.component';
-import { TranslateServiceStub } from '../../../../../test/translateServiceStub';
-import { EventPubSubService } from '../../services/eventPubSub.service';
 
 function removeExtraSpaces(text: string) {
   return `${text}`.replace(/\s{2,}/g, '');
