+++ packages/pagination-component/src/index.spec.ts
@@ -1,0 +1,11 @@
+import * as entry from './index';
+
+describe('Testing library entry point', () => {
+  it('should have an index entry point defined', () => {
+    expect(entry).toBeTruthy();
+  });
+
+  it('should have all exported object defined', () => {
+    expect(typeof entry.SlickPaginationComponent).toBe('function');
+  });
+});
