+++ packages/pagination-component/src/index.ts
@@ -1,0 +1,1 @@
+export * from './slick-pagination.component';
