+++ packages/pagination-component/src/slick-pagination.component.html
