+++ packages/pagination-component/src/slick-pagination.component.ts
