+++ packages/vanilla-bundle/package.json
@@ -51,6 +51,7 @@
     "@slickgrid-universal/custom-footer-component": "^0.14.1",
     "@slickgrid-universal/empty-warning-component": "^0.14.1",
     "@slickgrid-universal/excel-export": "^0.14.1",
+    "@slickgrid-universal/pagination-component": "^0.14.1",
     "@slickgrid-universal/text-export": "^0.14.1",
     "whatwg-fetch": "^3.6.2"
   },
